==> Step 0 - ContactSearchResults_aggregator.ipynb <==
import os
import pandas as pd
import re
import numpy as np

# Change filepath based on where you downloaded the contact search results to
contact_files = os.listdir(path = 'C:\\Users\\sgphsg\\OneDrive - FWD Group Management Holdings Limited\\Desktop\\ContactSearchResults')
contact_files = [i for i in contact_files if re.search('[0-9]',i)] # Remove aggregated excel files from previous runs

for index, file in enumerate(contact_files):
    if index == 0: # First file
        filepath =  'C:\\Users\\sgphsg\\OneDrive - FWD Group Management Holdings Limited\\Desktop\\ContactSearchResults' + '\\' + file
        df = pd.read_csv(filepath)
        month = re.findall('2021([0-9]{2})[0-9]{2}.csv$',file)[0]
        df['month'] = month
        print('Appending month {} to {} '.format(month, file))
    else:
        filepath =  'C:\\Users\\sgphsg\\OneDrive - FWD Group Management Holdings Limited\\Desktop\\ContactSearchResults' + '\\' + file
        df2 = pd.read_csv(filepath)
        month = re.findall('2021([0-9]{2})[0-9]{2}.csv$',file)[0]
        print('Appending month {} to {} '.format(month, file))
        df2['month'] = month
        df = pd.concat([df, df2], axis = 0)

df.head()

df.to_csv('C:\\Users\\sgphsg\\OneDrive - FWD Group Management Holdings Limited\\Desktop\\ContactSearchResults\\{}'.format('ContactSearchResults_agg.csv'),
          index = False)
