==> Step 1 - match_interaction_merge_dataset.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
import datetime as dt
import re
import joblib

import warnings
from pandas.core.common import SettingWithCopyWarning

warnings.simplefilter(action = 'ignore', category = SettingWithCopyWarning)
warnings.filterwarnings(action = 'ignore', category = DeprecationWarning)
warnings.filterwarnings(action = 'ignore', category = FutureWarning)


import boto3

META_RAW_FILEPATH = "s3://fwd-sg-sagemaker-raw-data/voice_analytics"
CRM_RAW_FILEPATH = "s3://fwd-sg-sagemaker-raw-data/call_center"

### Load data extracted from Step - 0 for merging into dataset ###

# Load Call records extracted from Amazon Connect
metadata = pd.read_csv('{}/ContactSearchResults_agg.csv'.format(META_RAW_FILEPATH))


# Load extracted CRM data from Step 0  
crmdata = pd.read_csv('{}/dl_crm.Filterednew_calllog_extraction000'.format(CRM_RAW_FILEPATH))

## Fuzzy Match Preparation for AMZ Connect Dataset

# The Amazon Connect Dataset needs to be filtered and transformed for merging.

# Remove calls not connected to agent
metadata = metadata[metadata['Agent'].notna()]

# Remove calls that are likely to be automated messages 
# Eg: d4a6bd73-a54b-4a8f-bee4-e0bff4c96fc7 is an example with "38s duration" 
metadata = metadata[pd.to_timedelta(metadata['Contact duration']) > dt.timedelta(seconds=45)]
metadata.shape[0]

# Clean contact number
metadata['Customer phone number'] = metadata['Customer phone number'].str[2:]
metadata['Customer phone number'] = metadata['Customer phone number'].str.replace('.00','',regex=False)


# Turn timestamp to datetime, and get epoch timestamp
metadata['Initiation timestamp'] = pd.to_datetime(metadata['Initiation timestamp'])
metadata['ts'] = metadata['Initiation timestamp'].values.astype(np.int64) // 10 ** 9
metadata['ts'] = metadata['ts'].apply(lambda x: np.nan if x < 0 else x)

metadata['Disconnect timestamp'] = pd.to_datetime(metadata['Disconnect timestamp'])
metadata['ts_end'] = metadata['Disconnect timestamp'].values.astype(np.int64) // 10 ** 9
metadata['ts_end'] = metadata['ts_end'].apply(lambda x: np.nan if x < 0 else x)

## Fuzzy Match Preparation for CRM Dataset

# The CRM Dataset needs to be filtered and transformed for merging.

# Turn timestamp to datetime, and get epoch timestamp
crmdata['interaction_date'] = pd.to_datetime(crmdata['call_time'])
crmdata['ts'] = crmdata['interaction_date'].values.astype(np.int64) // 10 ** 9
crmdata['ts'] = crmdata['ts'].apply(lambda x: np.nan if x < 0 else x)

# Get Calls classified under "Self-Help - Assist" only.
crmdata = crmdata[crmdata['subcategory1'].isin(['Self Help - Assist'])]

# Fuzzy Matching Function 

Connect function is to be applied on AMZ Connect Metadata Table. It fuzzy matches the connect metadata table and interaction table by querying for a match in the interaction table for every connect metadata row with 3 conditions using pandas apply function and adding the interaction_id of the matched to the connect metadata table for merging (inter_id).

The 3 conditions are as follows:

1. Same Customer Number (if multiple interactions for cust no. match to the closest start time)
2. Start time of tables within a specified time (RFE) AND Same agent (multiple match possible, in that case count as no match as put string in inter_id column)
3. End time of tables within a specified time (RFE) AND Same agent (multiple match possible, in that case count as no match as put string in inter_id column)

The "rfe" variable can be optimised to find the value with the most amount of matches (too high many multiple match, too low too little matches) 

NOTE: Accuracy of matches with conditions 2 and 3 are also not guaranteed, especially with higher "rfe" values

### Remove interactions which are non-phone based (email based), fro the purposes of analysis' ###

print('Self_assist interactions: {}'.format(len(crmdata)))
crmdata = crmdata[crmdata['customer_phone'].apply(lambda x: re.search('^[0-9]*$', str(x))).notnull()]
print('Self_assist calls: {}'.format(len(crmdata)))
# Deduplicate the crmdata - deduplication based on phone number and call time. 
crmdata = crmdata.drop_duplicates(['customer_phone','call_time'])
print('Self_assist calls deduped: {}'.format(len(crmdata)))
# Only accept inbound calls as per Lina's suggestion
crmdata = crmdata[crmdata['call_direction'] == 'Inbound']
print('Self_assist calls inbound: {}'.format(len(crmdata)))

# Cleaning for metadata - the phone numbers have one extra number
metadata['Customer phone number2'] = metadata['Customer phone number'].apply(lambda x: x[1:])

nb_matches = len(set(metadata['Customer phone number2']).intersection(crmdata['customer_phone']))
perc_matches = nb_matches / len(crmdata)

print('Number of phone number matches: {}'.format(nb_matches))
print('Percentage of phone numbers out of crm logs matched: {}'.format(perc_matches))

### Create join keys and merging both dataset ###

# Map agent to email as joining key between AMZ Connect and CRM datasets
agent_email_mapping_dict = {
    'Adeline Wong Sook Kuan': 'adeline.wong@fwd.com',
    'Bobby Teng Hao Han': 'bobby.teng@fwd.com',
    'CRM Administrator': None,
    'Donna Ang Lay Kheng': 'donna.ang@fwd.com',
    'Eric Jiang Hailong': 'eric.jiang@fwd.com',
    'Ethan Lua Wei Jun': 'ethan.lua@fwd.com',
    'Eugene Jacob Chan Dong Hao': 'eugene.chan@fwd.com',
    'Evon Thoo Yee Wan': 'evon.thoo@fwd.com',
    'Gavin Poon Joun Wae': 'gavin.poon@fwd.com',
    'Go Irish Dianne Vinaviles': 'irish.go@fwd.com',
    'Grace Chong Wei Hui': 'grace.chong@fwd.com',
    'Idio Kristine Avril Bebe': 'kristine.idio@fwd.com',
    'Jayde Ng Shu Ying': 'jayde.ng@fwd.com',
    'Kasturi Konasegaran': 'Kasturi.Konasegaran@fwd.com',
    'Kokilaletchmi D/O Govindasamy': 'koki.govindasamy@fwd.com',
    'Lim Yu Yan': 'yuyan.lim@fwd.com',
    'Marissa Binte Ismail': 'marissa.ismail@fwd.com',
    'Mooralli Raj S/O Sumoo Amatharlingam': None,
    'Muhammad Ridzwan Bin Yusri' : 'ridzwan.yusri@fwd.com',
    'Ngieng Sue Min': 'suemin.ngieng@fwd.com',
    'Noraini Taraman': 'noraini.taraman@fwd.com',
    'Nur Azlina Binte Abdul Rahman' : 'nurazlina.rahman@fwd.com',
    'Nurwati Binte Jamil': 'nurwati.jamil@fwd.com',
    'Pamella Krystle Bonayon Flores': 'pamella.flores@fwd.com',
    'Resie Arjonillo': 'resie.arjonillo@fwd.com',
    'Tracy Hwang Poh Sim': 'tracy.hwang@fwd.com',
    'Villegas Karl Geoff Duques': 'karl.villegas@fwd.com',         
}

crmdata['agent_email'] = crmdata['agent_name'].map(agent_email_mapping_dict)

# Create join keys for dataset
crmdata['join_key'] = crmdata['customer_phone'].astype(str) + crmdata['agent_email']
metadata['join_key'] = metadata['Customer phone number2'].astype(str) + metadata['Agent']

# Join the data
joined_data = pd.merge(crmdata, metadata, how = 'inner', left_on = 'join_key', right_on = 'join_key')

### Filter and match data based on timestamp ###

# Match the call_time (crmdata) and initiation_timestamp (metadata), or else row is thrown away
joined_data['call_time'] = pd.to_datetime(joined_data['call_time'])
joined_data['call_time_lower'] = joined_data['call_time'] - pd.Timedelta(minutes = 2)
joined_data['call_time_upper'] = joined_data['call_time'] + pd.Timedelta(minutes = 2)
joined_data = joined_data[((joined_data['Initiation timestamp'] <= joined_data['call_time_upper']) & 
                           (joined_data['Initiation timestamp'] >= joined_data['call_time_lower']))]

print('Proportion of matched data : {}'.format(len(joined_data) / len(crmdata)))
print('Number of calls matched: {}'.format(len(joined_data)))

# Save merged dataset into csv
joined_data.to_csv('./Data/merged_crm_meta.csv', index = False)

df = pd.read_csv('./Data/merged_crm_meta.csv')
df
